Clean up dead code
...after df9692c689089bb127ef520adacf089781fc4419 "tdf#118314 Code clean up on
imestatuswindow" had removed the sole consumer of aText.

(It is unclear to me whether the use of StatusDrawCallback in the
SalI18N_InputContext in vcl/unx/generic/app/i18n_ic.cxx is still necessary when
StatusDrawCallback does not do anything.  But StatusStartCallback and
StatusDoneCallback do not do anything, either, and are used exactly the same way
there, so lets keep all that, at least for now.)

Change-Id: Ic2a61819c6bb8ed192231ca309f03b6c29768f29
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/94760
Tested-by: Jenkins
Reviewed-by: Stephan Bergmann <sbergman@redhat.com>