tdf#131684 tdf#132236 sw_redlinehide: fix upper of frame moved...
... in SwUndoDelete::UndoImpl; this wasn't fixed properly in commit
4990de6d71163ff7bd756b79a76ae347ffa1a828 - the SwTextFrame was moved
into the SwSectionFrame but its upper was whatever it was previously.

This is rather ugly because with the constraint from the very special
case of tdf#131684 the MakeFrames() requires one node with pre-existing
frames, and it's not possible to move the pre-existing frame after
MakeFrames() because then the tdf#132236 case will have an empty
SwSectionFrame and InsertCnt_() will just delete it.

So try to detect the situation with some hack in InsertCnt_() to move
it into the SwSectionFrame.

(regression from c92941613e61624010befde4f8f772d323848a24)

Change-Id: Ic0199c85d7523a49676ad3df1d2c4d8fe135c2ff
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/94881
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@cib.de>