tdf#128502: Remove assertions that fire for gtktiledviewer on two docs
If you run gtktiledviewer so that it opens a spreadsheet and a text
document these assertions would fire.

In each case the assertion is followed by code that just returns if
the asserted condition doesn't hold, so it is unlikely that this
change will break anything.

There sure will be a lot of further changes needed to make such a use
case work. (For starters, editing either of the two documents does not
work.) But let's do it step by step.

Change-Id: I4063dbbcac7c77b57e1f4cb3f05bd5096c08c0fe
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/95037
Tested-by: Jenkins
Reviewed-by: Tor Lillqvist <tml@collabora.com>