tdf#133498 Make native drawing of FormattedField work (again)
Commit 676d8f737bbe8458580b94e94e2817e0b547b1b9
("lok: fix the window type of the formatted field control",
2020-05-06) had introduced the new 'WindowType::FORMATTEDFIELD'
and sets it for the 'FormattedField' control, instead of using
the previous default value set in the 'SpinField' ctor,
which is 'WindowType::SPINFIELD'.

This commit adds the corresponding 'WindowType::FORMATTEDFIELD'
cases relevant to make native drawing of the control work (again).

Also map the corresponding accessibility roles.

Change-Id: Ied0b0548c03dee3feaf1b6252c81b2a33b68b938
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/95156
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>