sd signature line: add initial menu item / UNO command
Signature lines already work in Writer and Calc, this starts adding it
to Draw.

The current code is similar to SID_DRAW_RECT, just the implementation
for the UNO command is allowed in read-only mode, to be used for PDFs
opened for signing.

Also allow SID_OBJECT_SELECT is read-only mode: selecting objects makes
sense to allow (usually already works on the UI) and this is needed to
allow finishing the "create" mode in a read-only view.

Change-Id: Icccb9ae1cd58eff1a7c2c36df1548ea7b412c246
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/95505
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins