Related: tdf#113278 PDF export: improve selection of signing certificate
The action should be "select", not "sign", since we just select the
certificate in the dialog. Signing will only happen later after the pdf
export is almost done.

Ignore gpg certificates which are not handled during pdf export. This
does not implement the request in the bug, but at least hides
certificates which don't work.

Change-Id: Ib12f2e38bf864b08838e5970cebdba9ff20603bc
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/95573
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins