tdf#132596 writerfilter: fix paste of RTF with header/footer
The problem is that writerfilter inserts the content of header/footer
into the body text, but then DomainMapper_Impl::PopPageHeaderFooter()
calls RemoveLastParagraph() and deletes a body paragraph containing a
fieldmark, and then in Undo some SwHistoryTextFieldmark can't find it,
and since 8786424257ad76c6656c1d2ef86e50d84a52a88e this crashes.

This is because of the borked error handling in
DomainMapper_Impl::PushPageHeaderFooter(); the
m_xBodyText->createTextCursorByRange() call throws an exception that is
swallowed, so the m_aTextAppendStack doesn't have an entry containing
the position in the header.

To fix the error handling, set m_bDiscardHeaderFooter = false only when
everything was successful.

Also fix the call to be xText->createTextCursorByRange instead
(this is a regression from 86afb613817e4ef1895f2a2e90d570403c8336cb).

Then it turns out that Undo still crashes, because sw can't Undo
changes of header/footer content, so just return early unless it's
a new document.

Change-Id: Ie5aeb45447c5fbd4b3ae15c4cffb9800583a6d1d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/95797
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@cib.de>