PPTX import: handle adjust values from both the shape and its placeholder
The direct problem is a crash in CustomShapeProperties::pushToPropSet(),
the code just hoped that the input file doesn't have more adjust values
than the # of adjust values we allocate based on the preset type. Fix
the crash, but there is a deeper problem here...

The shape can inherit custom shape properties from a placeholder, then
later it can have its own custom shape properties. When it comes to
adjust values specifically, we used to just append own adjust values to
the end of the list. This way we got the double of expected adjust
values. And later rendering took the N expected adjust values from
the start of the 2N element list, so we used the adjust values of the
placeholder, not of the actual shape.

Fix this by clearing the custom shape geometry once we know we have our
own preset geometry.

Change-Id: I09f669bf59c33b552b906733d323eba7af5548e7
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/95993
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins