tdf#133858 opening spreadsheet with large array formula takes 10 mins
the spreadsheet has lots of SUM(IF over a whole column.

Which results in us allocating a matrix with 1 million rows, which is
rather slow to process.

So reduce the matrix to the data that is actually there.

We can only do this for some opcodes, because other opcodes act
differently when referencing empty space, so for now I only perform this
optimisation for the opcode in use in this spreadsheet.

This takes the load time from 5m to 3s on my machine.

Change-Id: I41fe9afcb0fbdf2a928a19c44a0f291a1247a41c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/96022
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>