PPTX export, custom shape, bitmap fill: fix source vs fill rect confusion
Commit 8f56f83007659b34e3f272eab76e1cd6acb8dc1e (drawingML
export\import: cropping of shape's fill texture, 2014-02-12) improved
the DOCX filter, so the fill rectangle of a custom shape with bitmap
fill is handled.

The problem is drawingML has a source rectangle (similar to our crop
rect) to limit the usage of the bitmap, and also it has a fill rectangle
in case some margin is wanted around a stretched bitmap. We don't have a
mapping for the later.

Fix the problem by limiting the above work for DOCX, this way PPTX's
source rectangle won't be turned into a stretch's fill rectangle.

This way no unwanted margins will appear around the image -- those
margins can be large enough that the image effectively disappears on
export.

Change-Id: Ic35063545a56eec9eaf885bbd397a854705d134f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/96025
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins