crashtesting: sw: fix export of ooo24576-1.doc and ooo79410-1.sxw to odt
Crashes because an at-char fly has its anchor node deleted, and during
deletion of its text frame its SwAnchoredObject is updated, which
asserts because of inconsistent anchor node in the fly and
mpAnchorFrame; the immediate cause of the inconsistency is that
SwNodes::RemoveNode() changed the fly's anchor node.

The root cause is that in the sw_JoinText(bJoinPrev=true) code, a fly
anchored at the end of the deleted node isn't moved to the surviving
node.

SwTextNode::JoinPrev() uses different arguments to
ContentIdxStore::Save(), so use the same here.

The implementation of several ContentIdxStore functions, including
ContentIdxStoreImpl::SaveFlys(), ignore positions that are equal to the
passed nContent index, so passing in SwTextNode::Len() looks wrong.

(crash is regression from c2c1230522c195ed701f354962622685351c4962)

Change-Id: I3a4d54258611da6b15223273a187c39770caa8e2
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/93583
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@cib.de>