tdf#132321 sw: adapt fly at-para deletion to at-char wrt. sections
aecb5fabbfd505fd291e8e90e7c68f26ea7d21bb changed at-char selection to
check that not only the start is at the start of the section or the
end is at the end of the section, but that both are true.

Let's do this for at-para flys too for consistency, changing what was
introduced with 51c97e5e3ae88c43e611a4e6a61e375c66d5d007.

Change-Id: I1ec93b076d729288ce0809d1cfc24379aa9591ab
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/96125
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@cib.de>