tdf#133605: docx import: use ilvl value for list level import
Previous implementation was not aware about ilvl value which
declares a level number it is defining. So LO was able correctly
handle only lists were levels are defined in increasing order
starting from zero. Same problem was also with override lists: if
we just override only one some random level (for example ilvl=2
in bugdoc) it was applied to level 0.

Change-Id: If1f466fcbae0c82a8601e5b638e541d75aab665c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/96268
Tested-by: Jenkins
Reviewed-by: Thorsten Behrens <Thorsten.Behrens@CIB.de>