tdf#133990 sw_redlinehide: fix Undo with section before table
... at start of document.

The section node wasn't included in the PaM passed to SwUndoDelete,
so on Undo it called MakeFrames() starting from the contained table
node and crashed when reaching the section's end node.

Fixed, then it crashes in Redo, because SwDoc::DelSectionFormat()
calls SetModified() which calls some event listener which causes
EndAllAction() to be called, which by itself is scary and then it
crashes because the nodes array contains a bunch of deleted nodes.

Not sure if that SetModified() serves any purpose, but it's pointless
in Undo.

(1st crash is regression from 34c3b7af8c4f1c14529ca1e6f9a52f63fe0d8cff)

Change-Id: Iafdd073ca9577bf54dd5a8ad2eb8f8f110db93b6
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/96620
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@cib.de>