tdf#132326 sw_redlinehide: no frames allowed in Undo SwNodesArray
This crashes in SwAccessibleParagraph::GetRealHeadingLevel(); the a11y
should have been disposed earlier because the node it's connected to is
in the Undo SwNodesArray.

When SwUndoDelete::SwUndoDelete() calls SwNodes::MoveNodes(), all
existing frames must be destroyed; the MoveNodes() would reset its flag
temporarily inside a section, to delay creating the frames inside the
section when moving to the live SwNodesArray, but this also prevents
the deletion when moving in the other direction.

(crash is regression from c92941613e61624010befde4f8f772d323848a24)

Change-Id: I13719bf98ea96b8d68db3ee78cf7f4b61a99e7e5
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/96734
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@cib.de>