reset all draw gridOffsets when zoom needs to change...
so that these are recalculated on render, else the shape positions in
the tiles will be wrong. It is not enough to reset for just the objects
in the tile because the render code iterates over all objects and then
decide whether to render them based on the position/size.

An obvious optimization is to choose a view/window with its zoom
matching the tile-request using some useful heuristic and avoid a linear
scan of views.

Change-Id: Iffa96d261bc3362af3a528fbf7d4d3bcafc6d497
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/98167
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Dennis Francis <dennis.francis@collabora.com>
(cherry picked from commit 22d458164ae903f502e82f471c22ff8e1ebbb839)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/98162
Tested-by: Jenkins