sd: send annotation position for LOKit, fix multiple annotations
This adds support to send annotation position in the document when
sending the available annotations using LOKit.

It also fixes an issue with the json structure for the annotations,
which sends multiple annotation objects with the same (empty) key,
that after parsing reduces to only one (so only one of the many
annotations is present). This fix changes that so annotation
objects have each its own unique key.

Change-Id: I9d994383d6ee322f27f1426be5bd96ac81e8e609
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/97329
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>