tdf#134099 sw: fix textbox anchors on copy-paste and undo
Regression from commit 0df7a9f7c2fdaad35b8f35ca808edfcfb3ac5dc9
(fdo#82191 sw::DocumentLayoutManager: copy textbox content of draw
formats, 2014-08-15), without which this problem gets hidden, as copy
breaks the textbox into 2 pieces, so the textbox codepaths are no longer
hit.

The direct problem is that SwHistorySetFormat::SetInDoc() uses a raw
node index into the nodes array, which is past the end of the nodes
array.

Root cause is that we have this invariant that actions and their undo
has to be in sync, otherwise raw node indexes no longer work. In this
case, SwUndoSaveContent::DelContentIndex() did not delete a fly frame
format, because it was out of range, as it had a wrong anchor.

Fix this in SwTextFlyCnt::SetAnchor(), so that whenever the anchor of a
draw format is set via that function, we update its textbox as well.

Also fix a related problem when fly formats were copied twice.

Change-Id: I0d6c9069544c405eb20c5fed65fb40423b0adc84
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/97457
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>