clarify what WantMouseEvent is for
and add tristate machine support that we also need

WantMouseEvent is if the first click in a browse cell that activates it should
be passed to the control that appears on click as if it was itself clicked on

primarily for the the CheckBoxController to toggle it immediately.
Rework to explicitly toggle the checkbox in that implementation rather than
rely on passing fake mouse click/release/tracking events.

seeing as the main dbaccess table design view doesn't auto-launch its listboxes
on clicking in a cell, but the sub create index dialog does, drop the auto-launch
in the index dialog to match the main table design rathern than add a feature
to weld::ComboBox to auto-launch the popup for that one solitary case

Change-Id: Ie3d3f2ecf55d3d5b0b02b85ca09a6ca64bb800e9
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/98603
Tested-by: Caolán McNamara <caolanm@redhat.com>
Reviewed-by: Caolán McNamara <caolanm@redhat.com>