sw: layout: fix missing invalidation of text frames in tables
... when the position of the SwTabFrame changes.

The table is initially formatted on page 1, where one of its cells
overlaps flys anchored in the footer, so the SwTextFrame in it
contains SwFlyPortions.

As the table doesn't fit on page 1, the SwTabFrame moves forward to page
2; lcl_RecalcTable() is called a bit later to invalidate pos and size of
everything in the table.

However, it turns out that that's not enough, when SwTextFrame::Format()
is called it doesn't do anything because no part of the text has
actually been invalidated via InvalidateRange_().

If the SwTextFrame were moved on its own (not via table), then
SwContentFrame::MakeAll() would call
Prepare(PrepareHint::FramePositionChanged) which calls ClearPara().

The SwTabFrame is moved via SwFlowFrame::PasteTree(), which calls
SwTextFrame::Init() if it moves a text frame directly but does
nothing for tables.

So let's try to fix this similar to commit
5b273e2ba1a1ec6ae4fcc9b6b749db25d9cb92e0 by calling
Prepare(PrepareHint::FramePositionChanged) on every moved text frame
if the position actually changes, like SwContentFrame::MakeAll() does;
not sure what performance impact this has.

(apparently regression from 722d5186ba6253ef779cbca87535273bb3d352b8
 - whatever that means in this case, no idea how it worked before)

Note: the problem only reproduces on libreoffice-6-3 branch because
      libreoffice-6-4 and later have another layout change from commit
      c1b1049071a70f2d9c9f03961157096f8ae24f47 that needs reverting

Change-Id: I65d3e367d56b8799e1ed32172fbbc0249c2852eb
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/98925
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@cib.de>