Fix the minimal build-tools target
The revert commits change the build-tools target for a DESKTOP
build to build the complete LO. This restores the original,
minimal one and also adds a whitelist of allowd build types.

OpenCL needs a configure switch, as it's status is also stored
in a config header, so preventing the build is not enough.

This also reverts:
- commit 3fdda2e0e88a109f125c76919f9db40c43e364d9
- commit 5847a7019819ee73f1eb03e900f2119c561e9719

Change-Id: Ibfcb0c54e72da1b7c2e63c082ea6586520a787fa
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/102480
Tested-by: Jenkins
Reviewed-by: Jan-Marek Glogowski <glogow@fbihome.de>