lcms2: fix Windows Arm64 build
Adds the ARM64 platform to the VC2019 MSBuild solution. This turned
out to be tricky, because - for whatever reason - diff couldn't
detect the solution file as text, so I added the whole file to
replace it. And "-a" would just generate a UTF-16 diff, which is
not applyable, so this just adds the new file and copies it.

Change-Id: I538610bed071fd45dc107f405056f23e1bff7a09
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/102855
Tested-by: Jenkins
Reviewed-by: Jan-Marek Glogowski <glogow@fbihome.de>