fix SAL_WARN when instantiating Thesaurus
Before
    commit 59adbfa3ef4c75dbddeb70aa5615043dc5c0eac9
    lingucomponent: create instances with uno constructors
Thesaurus was constructed but Thesaurus::initialize was NOT being called
from GetAvailLocales.

However, we have some code marked "HACK" in
   cppuhelper/source/servicemanager.cxx
i.e.
   ServiceManager....::createInstanceWithArguments
which DOES call initialize.

And we have code in GetAvailLocales which passes what Thesuarus
considers an invalid number of arguments, because all the other similar
services DO take 2 args (even though they don't use the second arg!)

So we have a bunch of temporary hack and backwards compat stuff
interacting, causing a SAL_WARN.

So make Thesauras::initialize a little more tolerant, which means we're
now initialising it, which is a change in behaviour, so if this commit
comes up a regression, we will need find another warn to stop the
warning.

Change-Id: I60e0ed1bd2d4fbccbc539b5536aba055a7dfc6c4
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/99039
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>