tdf#134400 RTF import: fix unexpected inner properties for picture-in-textframe
Regression from commit 996e34f510fa975b80e869486a21d58dbce88d36 (RTF
import: implement support for tables inside text frames, 2013-06-16),
the problem was that both the outer "textbox" and the inner "picture
frame" object had a shapeType property, and the properties were stored
in a vector. So by the time RTFSdrImport::initShape() looked up the
shape type for the inner shape, it thought it's not a picture frame,
leading to data loss.

Change-Id: I4a536789371619d1d54afa8c8d41c7d273b0d21b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/99111
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>